site_160/requirements.txt
social-auth-app-django
mongoengine==0.13.0
django-bootstrap-form